Adding cleartool GtagsDB generation TAGS script and EXE
To Generate gtags simply gtags . for most of the case, but for
clearcase, it requires more works to properly generate the Gtags,
therefore the script is added.

EXE bundle some exe from cygwin which that needed to bundle with the exe
compiled from VB project

Bookmark.ini and GlobalBrowse.ini also needed by the application.